Script to ID problems in Lascar file naming
Script that pulls all files in the Main_study_exposure_assessment
directory and filters out files with certain naming problems. Saves the
problem files to a spreadsheet.

Here's what the spreadsheet contains:
The first two columns are "mstudyid" (for mother's studyid) and
"cstudyid" (for child's study ID, if applicable). Both of these are
identified by the first time the pattern  "BM....M" or "BM....C" occurs
in the file path.
If there is something in the "problem_id" column, this means that the
filepath contains a study ID name that matches neither the mstudyid nor
the cstudyid (if applicable).
If there is something in the "session_id" column, this means that the
filepath contains 2 or more session IDs, and there is no child ID in
the path (2 session IDs are allowed if there is both a mother and child
ID in the path).
If the value in the "duplicated" column is TRUE, this means the same
file (e.g. CU_C0_55_BM0304M_28Dec13_s_03_pri.txt) occurs more than once
in the "Main_study_exposure_assessment" directory (sometimes the file
might just be in the same folder twice, but sometimes the same file
appears in different folders and one of them should not be there.
NOTE: For these duplicated files, we want to make sure we don't
accidentally delete any actual data. I would suggest moving unnecessary
duplicate files to a new folder, such as "duplicate files", rather than
deleting them altogether. Also, please ensure that any duplicate files
actually are duplicates -- by opening the two files and checking that
the data matches (looking at the file start date and time, the monitor
name, and the first few values recorded by the monitor is a good
check).
The last 2 columns are the file name and the full file path of the
problem files.